+++ Terraria/GameContent/Generation/TrackGenerator.cs
@@ -1,8 +1,8 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.GameContent.Generation.TrackGenerator
-// Assembly: Terraria, Version=1.3.4.4, Culture=neutral, PublicKeyToken=null
-// MVID: DEE50102-BCC2-472F-987B-153E892583F1
-// Assembly location: E:\Steam\SteamApps\common\Terraria\Terraria.exe
+// Assembly: Terraria, Version=1.3.5.1, Culture=neutral, PublicKeyToken=null
+// MVID: DF0400F4-EE47-4864-BE80-932EDB02D8A6
+// Assembly location: F:\Steam\steamapps\common\Terraria\Terraria.exe
 
 using Microsoft.Xna.Framework;
 using System;
@@ -258,7 +258,7 @@ namespace Terraria.GameContent.Generation
 
     public static void Run(Point start)
     {
-      new TrackGenerator().FindPath(start.X, start.Y, 250, true);
+      new TrackGenerator().FindPath((int) start.X, (int) start.Y, 250, true);
     }
 
     private struct TrackHistory
